move inode-related stuff into inode.rs
